chore: ditching tauri for web-leptos demo
There's much more complexity than I hoped in using WASM/Leptos with
Tauri. For the demo's sake I am switching to web-leptos and call it a
day. After the demo and getting familiar with crux + leptos, I can take
the best path forward. will it be tauri? unlikely. probably time to get
on with swift after the demo.

this all demo thing is kind of a goal to learn more about crux in the
end. since the end goal of crux is native ui/ux, it would be great to
get those too and stop going circles with too many tech to learn at
once.

Signed-off-by: andreacfromtheapp <3269984+andreacfromtheapp@users.noreply.github.com>